fib 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,13 +6,13 @@
 */
 int main(void)
 {
-int num1, num2, next;
+unsigned long num1, num2, next;
 int i;
 
 num1 = 1;
 num2 = 2;
 next = num1 + num2;
-printf("%d, %d, %d, ", num1, num2, next);
+printf("%lu, %lu, %lu, ", num1, num2, next);
 for (i = 3; i < 99; i++)
 {
 num1 = num2;
@@ -20,11 +20,11 @@ num2 = next;
 next = num1 + num2;
 if (i == 98)
 {
-printf("%d", next);
+printf("%lu", next);
 }
 else
 {
-printf("%d, ", next);
+printf("%lu, ", next);
 }
 }
 printf("\n");
